feat(cli)!: V3 rename to autopg — structural (bin drop + package rename)
BREAKING. v3.0.0 rename to autopg. Structural changes only; doc rebranding
follows in a separate commit on this same branch.

Felipe directive 2026-05-10:
  - Q1: drop `pgserve` bin entirely; `autopg` is the sole CLI entry
  - Q4: package.json name `pgserve` → `autopg`
  - Q3 (keep): `pgserve_meta` postgres table untouched (transition step-by-step)
  - Q2 (LATER): repo rename `namastexlabs/pgserve` → `automagik-dev/autopg`
    is the final cutover step; NOT in this PR. Trust regex + install.sh
    REPO + GitHub URLs all still point at namastexlabs/pgserve.

Structural changes (24 files / 465 insertions / 475 deletions, net negative):

  bin/
    - REMOVED bin/autopg-wrapper.cjs (the 16-line delegator)
    - RENAMED bin/pgserve-wrapper.cjs → bin/autopg-wrapper.cjs via git mv
      (preserves blame on the 414-line dispatcher logic)
    - Header comment rewritten — autopg is now the only entry point;
      historical pgserve alias removed.

  package.json
    - "name": "pgserve" → "autopg"
    - bin: dropped "pgserve" entry; only "autopg" remains
    - Note: package-lock.json still has stale "pgserve" entries; lockfile
      is npm bookkeeping (we use bun.lock canonically). Harness denied
      lockfile edits; can fix-up if bot flags. v3 drops npm publish
      anyway (see PR #122), so package-lock drift has zero runtime impact.

  Internal file references updated (sed replace, 22 files):
    bin/postgres-server.js, knip.json, scripts/postinstall.cjs,
    scripts/test-bun-self-heal.sh, src/cli-config.cjs, src/cli-install.cjs,
    src/cli-restart.cjs, src/cli-ui.cjs, src/index.js,
    tests/cli-install.test.js, tests/cli/{config,ui,uninstall,verify}.test.js,
    tests/console/{auth,smoke}.test.js, tests/e2e/settings-flow.test.js,
    tests/integration/{gc-provision,v2.6-cohort-smoke,verify-slug-rotation}.{test.sh,test.sh},
    tests/postmaster-args.test.js

Preserved deliberately (per Felipe's "step-by-step" directive):
  - `pgserve_meta` postgres table — separate from autopg_meta; different
    lifecycle; renaming requires DDL migration on deployed hosts.
  - `namastexlabs/pgserve` GitHub URLs + install.sh REPO + trust regex —
    repo rename is the LAST cutover step.
  - `wish/autopg-distribution-cutover-finalize` and other wish slugs that
    contain "pgserve" — historical wish names.
  - `src/cli-install.cjs` filename (and other src/* filenames) — names
    aren't user-visible.
  - package-lock.json — npm artifact; bun.lock is canonical post-V3-2.

Validation
  - `bun test` — 684 pass / 3 skip / 8 fail. Same numbers as origin/main
    pre-rename (verified via git stash + retest on the prior V3-1 PR).
    The 8 failures are pre-existing infra tests (pgserve url/port/status/ui)
    NOT introduced by this rename.
  - `grep -rE "pgserve-wrapper\\.cjs |\"pgserve\":" --include='*.cjs' ...` —
    clean except: (a) the historical-annotation header comment in
    bin/autopg-wrapper.cjs noting the rename, (b) package-lock.json stale
    entries (intentionally left, see note above).

Doc rebranding (operator-facing `pgserve <verb>` → `autopg <verb>` across
docs, README, CHANGELOG, install.sh help text, wish files) lands in the
follow-up commit on this same branch.
