+++ knip.json
@@ -3,7 +3,7 @@
   "entry": [
     "src/index.js",
     "bin/postgres-server.js",
-    "bin/pgserve-wrapper.cjs",
+    "bin/autopg-wrapper.cjs",
     "src/update/index.js",
     "src/commands/**",
     "src/cosign/**",
