+++ src/index.js
@@ -4,7 +4,7 @@
  * Public surface after the `pgserve-singleton-no-proxy` Group 2 deletion:
  * the bun proxy data plane, daemon control socket, libpq protocol
  * rewriting, and SO_PEERCRED handshake are gone. Operators interact with
- * pgserve through the CLI (`bin/pgserve-wrapper.cjs`), the postmaster
+ * pgserve through the CLI (`bin/autopg-wrapper.cjs`), the postmaster
  * subcommand (`bin/postgres-server.js postmaster`), and the cohort-shared
  * helpers under `src/lib/`.
  *
