+++ tests/cli-install.test.js
@@ -21,7 +21,7 @@ import path from 'node:path';
 import { spawnSync } from 'node:child_process';
 
 const REPO_ROOT = path.resolve(__dirname, '..');
-const BIN = path.join(REPO_ROOT, 'bin', 'pgserve-wrapper.cjs');
+const BIN = path.join(REPO_ROOT, 'bin', 'autopg-wrapper.cjs');
 
 let tmpHome;
 let stubBin;
@@ -683,7 +683,7 @@ describe('pgserve singleton (v2.4) — socket dir + admin.json supervisor record
 describe('serve alias', () => {
   test('pgserve serve --bogus-flag re-routes to postmaster (singleton v2.4)', () => {
     // pgserve singleton (v2.4): the bun-proxy daemon is gone. `serve` now
-    // aliases to `postmaster` via bin/pgserve-wrapper.cjs. We can't fully
+    // aliases to `postmaster` via bin/autopg-wrapper.cjs. We can't fully
     // exercise the postmaster without postgres binaries on PATH, so we
     // just verify the wrapper proceeded past the install short-circuit
     // (stderr will be a bun probe error or a postmaster-mode error, never
